+++ clients/privacy-center/types/api/models/GCSAuthMethod.ts
@@ -1,0 +1,8 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+export enum GCSAuthMethod {
+  ADC = "adc", // Application Default Credentials
+  SERVICE_ACCOUNT_KEYS = "service_account_keys",
+}